New Raspbian build, fixes #1168 #1165

+++ src/build/installer.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 PMMP_VERSION=""
 MAC_BUILD="PHP_5.5.8_x86_MacOS"
-RPI_BUILD="PHP_5.5.3_ARM_Raspbian_hard"
+RPI_BUILD="PHP_5.5.8_ARM_Raspbian_hard"
 AND_BUILD="PHP_5.5.8_ARMv7_Android"
 update=off
 
@@ -69,26 +69,30 @@ if [ $update == on ]; then
 else
 	echo -n "[3/3] Obtaining PHP:"
 	echo " detecting if build is available..."
-	if [ "$(uname -s)" == "Darwin" ]; then	
-		rm -r -f bin/ >> /dev/null 2>&1
-		echo -n "[3/3] Mac OSX PHP build available, downloading $MAC_BUILD.tar.gz..."
-		download_file "http://sourceforge.net/projects/pocketmine/files/builds/$MAC_BUILD.tar.gz" | tar -zx > /dev/null 2>&1
-		chmod +x ./bin/php5/bin/*
-		echo -n " regenerating php.ini..."
-		OPCACHE_PATH=$(find "./bin/php5" -name opcache.so)
-		echo "zend_extension=\"$OPCACHE_PATH\"" > "./bin/php5/lib/php.ini"
-		echo "opcache.enable=1" >> "./bin/php5/lib/php.ini"
-		echo "opcache.enable_cli=1" >> "./bin/php5/lib/php.ini"
-		echo "opcache.save_comments=0" >> "./bin/php5/lib/php.ini"
-		echo "opcache.fast_shutdown=1" >> "./bin/php5/lib/php.ini"
-		echo "opcache.max_accelerated_files=4096" >> "./bin/php5/lib/php.ini"
-		echo "opcache.interned_strings_buffer=8" >> "./bin/php5/lib/php.ini"
-		echo "opcache.memory_consumption=128" >> "./bin/php5/lib/php.ini"
-		echo "opcache.optimization_level=0xffffffff" >> "./bin/php5/lib/php.ini"
-		echo "date.timezone=$TIMEZONE" >> "./bin/php5/lib/php.ini"
-		echo "short_open_tag=0" >> "./bin/php5/lib/php.ini"
-		echo "asp_tags=0" >> "./bin/php5/lib/php.ini"
-		echo " done"
+	if [ "$(uname -s)" == "Darwin" ]; then
+		if ["$(uname -s)" == iPhone*] || ["$(uname -s)" == iPod*] || ["$(uname -s)" == iPad*]; then
+		
+		else
+			rm -r -f bin/ >> /dev/null 2>&1
+			echo -n "[3/3] Mac OSX PHP build available, downloading $MAC_BUILD.tar.gz..."
+			download_file "http://sourceforge.net/projects/pocketmine/files/builds/$MAC_BUILD.tar.gz" | tar -zx > /dev/null 2>&1
+			chmod +x ./bin/php5/bin/*
+			echo -n " regenerating php.ini..."
+			OPCACHE_PATH=$(find "./bin/php5" -name opcache.so)
+			echo "zend_extension=\"$OPCACHE_PATH\"" > "./bin/php5/lib/php.ini"
+			echo "opcache.enable=1" >> "./bin/php5/lib/php.ini"
+			echo "opcache.enable_cli=1" >> "./bin/php5/lib/php.ini"
+			echo "opcache.save_comments=0" >> "./bin/php5/lib/php.ini"
+			echo "opcache.fast_shutdown=1" >> "./bin/php5/lib/php.ini"
+			echo "opcache.max_accelerated_files=4096" >> "./bin/php5/lib/php.ini"
+			echo "opcache.interned_strings_buffer=8" >> "./bin/php5/lib/php.ini"
+			echo "opcache.memory_consumption=128" >> "./bin/php5/lib/php.ini"
+			echo "opcache.optimization_level=0xffffffff" >> "./bin/php5/lib/php.ini"
+			echo "date.timezone=$TIMEZONE" >> "./bin/php5/lib/php.ini"
+			echo "short_open_tag=0" >> "./bin/php5/lib/php.ini"
+			echo "asp_tags=0" >> "./bin/php5/lib/php.ini"
+			echo " done"
+		fi
 	else
 		set +e
 		grep -q BCM2708 /proc/cpuinfo > /dev/null 2>&1
